app/requirements.txt
flask
prometheus_client
pytest
requests
psycopg2-binary
